// src/types/suppliers.type.ts
export type SupplierType = {
  id: string;
  name: string;
  contact: string;
  email: string;
  GSTNumber: string;
  taxNumber: string;
  openingBalance: number;
  outStandingBalance: number;
  address: string;
  postCode: string;
  country: string;
  state: string;
  city: string;
  createdAt: string;
  updatedAt: string;
}; 